This function creates a row of a LaTeX table.

This function creates a row of a LaTeX table.

Usage

TexRow(
  value,
  cspan = rep(1, length(value)),
  position = "r",
  surround = "%s",
  space = 0,
  dec = 3,
  percentage = FALSE,
  dollar = FALSE,
  se = FALSE,
  pvalues = NULL
)

Arguments

value

The value(s) to be formatted. Must be a numeric or character vector.

cspan

(integer). If greater than 1, multicolumn{cspan}{position}{value} will be used. For example, cspan=c(1,2,1) means that the second entry of value should span 2 columns. Default is cspan = rep(1, length(value)).

position

(character). If cspan > 1, multicolumn{cspan}{position}{value} will be used. For example, position=c("l","c","r") means that the second entry of value should be centered. Default is "r".

surround

(character). This will be applied to the value as sprintf(surround, value), so surround must contain the "%s" placeholder. Default is "%s".

space

(numeric). The number of points (pt) of vertical space to append to the end of the row. Default is 0.

dec

(integer). Only relevant if value is numeric. Number of decimal places. If scalar, the same decimal will be used for each entry of value. If vector, must be the same length as value. Default is 3.

percentage

(logical). Only relevant if value is numeric. If TRUE, a percentage symbol "%" will be added to the end of each entry of value. If scalar, it will be used for all entries of value. If vector, must be the same length as value.

dollar

(logical). Only relevant if value is numeric. If TRUE, a dollar sign "$" will be added to the end of each entry of value. If scalar, it will be used for all entries of value. If vector, must be the same length as value.

se

(logical). Only relevant if value is numeric. If TRUE, value will be wrapped in parentheses. If scalar, it will be used for all entries of value. If vector, must be the same length as value.

pvalues

(numeric). Only relevant if value is numeric. If not NULL, must be numeric. If less than 0.1, a star will be added to the end of each entry of value. If less than 0.05, a second star will be appended. If less than 0.01, a third star will be appended. If scalar, the same p-value will be assumed for all entries of value. If vector, must be the same length as value.